+++ src/client/api.ts
@@ -4,16 +4,19 @@
 'use strict';
 
 import { Event, Uri } from 'vscode';
-import { NotebookCell } from 'vscode-proposed';
 import { isTestExecution } from './common/constants';
 import { traceError } from './common/logger';
 import { IConfigurationService, Resource } from './common/types';
-import { JupyterExtensionIntegration } from './datascience/api/jupyterIntegration';
-import { IDataViewerDataProvider, IDataViewerFactory } from './datascience/data-viewing/types';
-import { IJupyterUriProvider, IJupyterUriProviderRegistration, INotebookExtensibility } from './datascience/types';
 import { getDebugpyLauncherArgs, getDebugpyPackagePath } from './debugger/extension/adapter/remoteLaunchers';
 import { IInterpreterService } from './interpreter/contracts';
 import { IServiceContainer, IServiceManager } from './ioc/types';
+import { JupyterExtensionIntegration } from './jupyter/jupyterIntegration';
+import {
+    IDataViewerDataProvider,
+    IDataViewerFactory,
+    IJupyterUriProvider,
+    IJupyterUriProviderRegistration
+} from './jupyter/types';
 
 /*
  * Do not introduce any breaking changes to this API.
@@ -82,8 +85,6 @@ export interface IExtensionApi {
         };
     };
     datascience: {
-        readonly onKernelPostExecute: Event<NotebookCell>;
-        readonly onKernelRestart: Event<void>;
         /**
          * Launches Data Viewer component.
          * @param {IDataViewerDataProvider} dataProvider Instance that will be used by the Data Viewer component to fetch data.
@@ -106,7 +107,6 @@ export function buildApi(
 ): IExtensionApi {
     const configurationService = serviceContainer.get<IConfigurationService>(IConfigurationService);
     const interpreterService = serviceContainer.get<IInterpreterService>(IInterpreterService);
-    const notebookExtensibility = serviceContainer.get<INotebookExtensibility>(INotebookExtensibility);
     serviceManager.addSingleton<JupyterExtensionIntegration>(JupyterExtensionIntegration, JupyterExtensionIntegration);
     const jupyterIntegration = serviceContainer.get<JupyterExtensionIntegration>(JupyterExtensionIntegration);
     const api: IExtensionApi = {
@@ -152,9 +152,7 @@ export function buildApi(
                     IJupyterUriProviderRegistration
                 );
                 container.registerProvider(picker);
-            },
-            onKernelPostExecute: notebookExtensibility.onKernelPostExecute,
-            onKernelRestart: notebookExtensibility.onKernelRestart
+            }
         }
     };
